feat(ingest): add ACSC RSS adapter
